+++ jablib/src/main/java/module-info.java
@@ -3,13 +3,14 @@ open module org.jabref.jablib {
     exports org.jabref.logic;
 
     exports org.jabref.search;
-    exports org.jabref.migrations;
     exports org.jabref.logic.search;
     exports org.jabref.logic.search.query;
     exports org.jabref.model.entry.field;
     exports org.jabref.model.search;
     exports org.jabref.model.search.query;
     exports org.jabref.model.util;
+    exports org.jabref.logic.preferences;
+    exports org.jabref.logic.importer;
 
     requires javafx.base;
     requires com.tobiasdiez.easybind;