+++ jablib/src/main/java/org/jabref/logic/importer/OpenDatabase.java
@@ -2,14 +2,8 @@ package org.jabref.logic.importer;
 
 import java.io.IOException;
 import java.nio.file.Path;
-import java.util.Arrays;
-import java.util.List;
 
 import org.jabref.logic.importer.fileformat.BibtexImporter;
-import org.jabref.migrations.ConvertLegacyExplicitGroups;
-import org.jabref.migrations.ConvertMarkingToGroups;
-import org.jabref.migrations.PostOpenMigration;
-import org.jabref.migrations.SpecialFieldsToSeparateFields;
 import org.jabref.model.util.FileUpdateMonitor;
 
 public class OpenDatabase {
@@ -26,20 +20,6 @@ public class OpenDatabase {
     public static ParserResult loadDatabase(Path fileToOpen, ImportFormatPreferences importFormatPreferences, FileUpdateMonitor fileMonitor)
             throws IOException {
         ParserResult result = new BibtexImporter(importFormatPreferences, fileMonitor).importDatabase(fileToOpen);
-        performLoadDatabaseMigrations(result, importFormatPreferences.bibEntryPreferences().getKeywordSeparator());
         return result;
-    }
-
-    private static void performLoadDatabaseMigrations(ParserResult parserResult,
-                                                      Character keywordDelimited) {
-        List<PostOpenMigration> postOpenMigrations = Arrays.asList(
-                new ConvertLegacyExplicitGroups(),
-                new ConvertMarkingToGroups(),
-                new SpecialFieldsToSeparateFields(keywordDelimited)
-        );
-
-        for (PostOpenMigration migration : postOpenMigrations) {
-            migration.performMigration(parserResult);
-        }
     }
 }